Fix comparing strings with multibyte chars
Using `x.len()` on a str returns the number of bytes, which is not
always the same as the number of characters. For example, the characters
'ö' or '香' are representing using more than one byte. This leads
to strange behaviour (e.g. `levenshtein("a", "ä")` would return 0).
Replace calls to `str.len()` with `str.chars().count()`, which returns
the correct number of characters in a string.