levenshtein optimisation #2
use new split_on_common_prefix() helper

In cases of two non-equal strings with equal byte length, some portion of
them would be compared in the equality check, before the check fails and
it moved on to a more thorough calculation. Unfortunately, the information
about what length of the start of the strings did match was thrown away
(not made available from the basic equality check), which would have been
useful, since we could skip past it (the common prefix has zero value).

The new helper allows precisely this. It finds the common prefix, splitting
the strings after it. We can then ignore it, and proceed on with just the
suffixes that remain (if any).