levenshtein optimisation #3
The new 'normalised' form does char counting on top of the standard
algorithm it calls. This change avoids unnecessary recalculation by
moving the main algorithm to a private function which takes `Option`
wrapped counts, allowing the normalised form to pass in the values it
calculates, thus letting this be done only once.