j-w optimisation #1
This optimisation addresses the fact that in the Jaro-Winkler
implementation, any common prefix is processed to determine its
char count in addition to this portion of the stirngs also being
processed within the use of the standard Jaro algorithm.

Here we move the Jaro implementation to an 'inner' function, which
takes an additional parameter - the char count of a common prefix
that has been removed from the provided pair of strings. The
algorithm can then take this into account in calculating the metric.

The standard Jaro public API function calls this with both full
strings and a value of zero for this new param.

The Jaro-Winkler implementation now uses the prefix-splitting
helper, passing the pair of suffixes in to the 'inner' Jaro algorithm,
along with the char-count of the prefix. The adjustment is then
performed as before. This avoids wasting unnecessary effort on
the prefix portion within the use of the Jaro algorithm.