jaro optimisation #4 & j-w optimisation #3
Taking the j-w optimisations further, this makes use of the prefix
splitting helper within the inner Jaro algorithm.

The function has been modified such that instead of taking a char-count of
the size of the common prefix removed from the pair of strings, it now
optionally takes a pointer to return the count, obtaining it within the
function through use of the helper internally.

Using the prefix splitting helper within the function means that we avoid
doing a `.chars().count()` iteration over the prefix twice, once going
over `a` and once going over `b`. It also then allows the main part of the
algorithm to completely avoid processing the common prefix portion of the
strings.