Use a flat vector in Damerau-Levenshtein
Instead of representing a 2x2 grid with a vector of vectors, just use a single
vector to improve performance. We can do this since the dimensions are fixed.

This method was suggested by @lovasoa as an alternative to adding a dependency
on the ndarray crate.

In my benchmark testing, the new approach is about as fast using ndarray. On my
machine, the original approach takes about 22,000 ns/iter, whereas the new
approach takes about 17,000 ns/iter.

See https://github.com/dguo/strsim-rs/issues/34 for more context.